+++ backend/src/main/java/syb/moviepedia/movie/external/tmdb/dto/TmdbMovie.java
@@ -10,7 +10,7 @@ import java.util.List;
  */
 public record TmdbMovie(
         @JsonProperty("id")
-        Long code,
+        Integer code,
 
         String title,
